Add a thresh parameter to the smooth function so we can pass it in.
The smoothing was simply setting any differences between successive
values greater than one standard deviation to 0.0.

However, for pressure values some bad values can give rise to deltas
less than this.  The problem is that the bad values are driving up the
sd to to point where this test doesn't work.

Passing a simple float threshold of 5.0 seems to work well for the
altitude and temp values, but over-smooths the pressure values.

I am happy with these plots though.

Next job is to do some fourier analysis on these series.